Fixed some x64 global mapping bugs
Now the rewriter will rewrite the entirety of libc, but the local
lookup function seems to be broken still.  Somehow the entries are
not correct in the mapping of the library, while the entries
for the main binary's mapping are fine.  Therefore, full rewriting
still doesn't work yet, but the global mapping seems to work great
so far, as it successfully looked up the local lookup for libc.