Added debug text to check instruction lengths
It appears that now there is a difference in instruction lengths
between the mapping and code generation phases.  Therefore, the
mapping is probably incorrect, and it is being placed at the
wrong offset.  It is still unclear what is causing this, but I
need to switch to implementing a shadow stack because that is
higher priority.