Added simple shadow stack implementation
The shadow stack is implemented as an overwriting, parallel shadow
stack at a fixed offset from the actual stack.  The current way of
determining the location of the memory to allocate for the stack is
done by disabling aslr with the command:
	setarch `uname m` -R
The -R flag disables aslr.  Unfortunately, even with aslr disabled,
it appears the stack is sometimes allocated at a different address
depending on the binary, so the "right" way to do this would
probably be to allocate the shadow stack at runtime after figuring
out the stack's base address.  However, time right now is limited.

This shadow stack implementation unfortunately does not use the
instrumentation API.  One major problem with it is that it needs
access to the ORIGINAL addresses that are placed on the stack.
This only happens partway through a rewritten call instruction
(in the original binary, the call would push a return address and
jump to the destination; in many shadow stack implementations,
the instrumented code goes in each function prologue, AFTER the
call instruction, because the return address isn't on the stack
until after the call).  The inserted code that "pushes" the return
address on the shadow stack needs to be placed AFTER the code that
looks up the old return address and pushes it on the real stack,
but BEFORE the final jmp instruction that actually jumps to the
function that was called.  I could have an instrumentation type
that inserts instrumentation mid-instruction, but what that means
would not be clear, and I'm not certain in which cases it would
be useful.