Fixed instruction length mismatch bug
Since I am adding instructions from the binary to the assembler
cache, the instructions retrieved from the binary may be laden
with garbage prefixes.  This means that two instructions that
disassemble to the same instruction may not actually have the same
length in their original bytes passed to the disassembler.  Since
every time I encountered an instruction and had an unfinished
mapping I added the instruction to the cache, the size of a certain
instruction could fluctuate when a new version of it was added
to the cache.  This was solved by only adding the first version
of the instruction we come across.  This may increase the length
of some instructions by using a version with extra prefixes that
the original instruction didn't have, but this is hard to avoid
since some of these instructions do not appear to want to
reassemble, and therefore taking the original bytes with a replaced
rip offset is the only way to preserve them.

Unfortunately, I should have double-checked Intel's x86/64
reference manual, because immediate values, which can appear in
some instructions referencing rip, appear after the displacement
bytes in an instruction's encoding.  That means that my current
displacement rewriting technique does not work for any instructions
that contain an immediate value.  I have not fixed this yet, but
it appears that capstone provides enough semantic information that
this can be fixed without resorting to an even more jury-rigged
solution.