Fixed rip-replacing code to modify correct displacement bytes
Now rewriting mostly works, but there may be problems with some
instructions that change the length of the instruction retrieving
an offset from rip.  A simple test program runs, but crashes in
a call to __call_tls_dtors, when it calls _dl_find_dso_for_object.
I think that I don't currently recognize the dynamic linker (for
which I will need to rewrite the return address before jumping
to the linker), but it also appears that the plt call, which seems
to resolve to __tls_get_addr instead of _dl_find_dso_for_object,
does not get the right offset when rewritten---it's a few bytes
off.  This seems to me to be potentially two independent problems
that will need to be solved before the program can run to
completion, but it also requires some additional study.