Fixed rip displacement bug
By passing the length of the new instruction (including the length
of all instructions preceding and including the instruction
referencing rip), the displacement now appears to be correct for
all rewritten instructions.

This also fixes what would have been a problem for inserting
instrumentation code, since it would have broken displacements.

The rewritten binaries still break when attempting to enter the
dynamic linker because the global mapping currently has no way to
identify it, so it halts due to there being no entry for it.