Added handling of external address ranges to global mapping
Now the global mapping will rewrite a return address on the stack
if an entry in the global mapping contains 0x00000000.  This is
done for code in the binary that I cannot or do not rewrite.
Specifically, this is [vdso], [vsyscall], and the dynamic linker,
which I find by assuming its name starts with "/ld-" and that no
other library will contain that substring.  Matching this is a bit
of a hack right now, but works fine for now.

I also assume that any code entering these regions is the result of
a call instruction so that I can rewrite the return address. Any
region determined to be external (done by checking the lookup
address right before jumping to it---if it is zero, there is no
valid lookup and it is instead external) has the original address
returned, so control flows into the non-rewritten external region
and later returns to the rewritten return address.

However, right now there is a bug in which the recursive call to
the global lookup obliterates the rax value that was saved outside
the stack by certain rewritten instructions.  I need to adjust the
distance that rax is saved outside the stack so that it is not
overwritten by the expanding stack.  This has been a problem in
the past as well, because I keep making changes that push more
values on the stack and therefore make the stack grow more.
I could simply save it a large distance outside the stack but
I want to keep it as close as possible so that I don't get crashes
by accidentally writing outside the stack memory region.