Modified global mapping to prevent overwriting saved rax value
Before calling the local lookup, the value of rax is saved outside
the stack to avoid changing the stack pointer.  This was done
because after that, rax is used to save the target of an indirect
control flow instruction, which potentially can use the stack
pointer, so we needed to leave it alone to save rax.  In theory we
could save it just outside the stack and then decrement rsp right
after getting the jump target value, but that's not what I did
originally.

I chose a semi-arbitrary value for how far outside the stack to
save the value (64).  I increased it several times when originally
writing the 32-bit lookup functions.  When I changed the code over
for x86-64, I ended up pushing more onto the stack than I did in
x86.  I have now modified the global lookup to not expand the
stack as much, mostly by simply restoring as many saved registers
as possible before making another call.  I did also end up saving
rax inside the global mapping outside the stack to provide, I hope,
exactly the amount of space needed to not need to change the value
I originally chose for how far outside the stack to save rax.  If
I encounter problems after this point, I will probably change the
code to simply change the stack pointer after retrieving the jump
target, but for now it stays as it is.