Fixed minor bug with rewriting main executables
Since the rewriter replaced rip differently in main executables
from shared libraries, it replaced rip completely with an absolute
address for the main executable.  This meant that every instruction
containing rip had to be re-assembled by the assembler, and
unfortunately there are a number of instructions that the
assembler refuses to assemble.  The solution when replacing rip
for shared objects was to re-use the original disassembled bytes,
but this only works if the memory addressing mode doesn't change.

Fortunately, we can just use rip for the main binary as well and
use relative addressing like with shared objects.  This unifies
the replace_rip code for both the main binary and shared objects,
plus paves the way for support for position-independent main
executables in the future.