Added code to overwrite start of old .text section with phdrs
Now the phdrs are placed at the start of the .text section if there
is room for them there.  I still need to implement code to
restore the code bytes at program startup, but for now, all
programs I have tested that have __do_global_dtors_aux at a far
enough offset in the .text section seem to work.  If the program's
main function is not large enough, the phdrs will overwrite this
function, which causes a crash at the end of the program.  It
appears that none of the programs I have tested have inline data
in the first 700 or so bytes, so for most of them the clobbered
code at the start doesn't make a difference and they run without
any problem.

Right now I save a copy of the first page of bytes from .text. This
means that I assume the phdrs will not occupy more than 4096 bytes.
For restoring the code bytes, I have to make the first page of the
.text section writable, copy the first page that I have saved in a
separate section over, and then make it read-only again.