Added assembly to restore saved first page of overwritten .text
Now, if the phdrs can fit inside the text section, the code at
the start of the .text section is overwritten with the phdrs
and then restored at runtime.  This seems to not cause any
complaints from the linker, although I still have to do a bit more
testing to be sure.  I only restore the *remaining* bytes in the
first page after you skip over the old phdrs, plt, etc.  As long
as the phdrs aren't absolutely massive, this won't make a
difference, but if it does we may want to save 2 pages instead. I
do this because we have to change page permissions aligned to page
boundaries.

I changed the code determining whether .text was being overwritten,
and probably still could do some cleaning up of that code to get
rid of some of the code that resulted from the approach I was
following before.