Fixed bug with identifying callbacks in x86-64 binaries
The code for identifying callbacks in the PLT (for rewriting
only the main binary) was broken because I had changed the code
for replacing rip and called the get_remap_callbacks_code
function before replacing rip.  The original regex for
recognizing a rip-relative address assumed that rip had
been replaced by an address, but that is not how I ended
up replacing references to rip.  The new way I replace
rip makes it simpler to leave the call to
get_remap_callbacks_code where it is, but change the
regex to refer to the original operands and calculate
the final computed address differently.

I also found that behavior for rewritten 64-bit binaries
when using the no_pic option is suspicious.  I think
that the no_pic code may not be working correctly for
64-bit binaries.  I may decide to disable the option for
64-bit binaries until this is solved.

I also noticed that simplest.c, when compiled as a
64-bit binary with optimization -O0 *writes outside*
the stack.  The code was ridiculously inefficient and
broke the icount instrumentation code to boot.  This
means some binaries may require even more careful
instrumentation code.  Changing to optimization level
-O1 makes the binary work without issues.